refactor(collector): extract partition.go from base.go
Move partition management (weekly bounds, creation, GC) into a
dedicated file.

Part of #75.
